tests: move 'bookkeeper' centric tests to their own file
Should we pull these out into a separate test suite?
